requirements.txt
tensorflow==2.9.1
streamlit==1.12.0
numpy==1.23.1
keras==2.9.0
pillow==9.1.1
